staging: slhttp: implement very simplistic HTTP response
With this change, we can already aim haproxy and ab at the driver.
It saturates around 80kpps, but still achieves a higher connection
rate than haproxy on localhost (about +20%).

We had to emit an HTTP/1.0 + keep-alive response because of AB.
Note that we could very well choose 1.0/1.1 from the request since
we don't need this information later.

Signed-off-by: Willy Tarreau <w@1wt.eu>