staging: slhttp: build a complete response based on the request
Now we parse the request, respond in the same HTTP version as the
client, enable HTTP keep-alive if the request starts with "/k/",
otherwise close, and send the amount of bytes requested in the URI
(with possible rounding near the boundary of a packet size). The
padding is automatically computed and the final states are properly
handled (eg: we retransmit the FIN only if requested, and we send
the last ACK in response to a FIN).

Outgoing packets are filled with readable data if they're smaller
than 200 bytes (debugging), otherwise they're not touched.

Multiple packets still don't work though the packet count is already
present.

Preliminary tests show an about 45% improvement in connection rate vs
httpterm when the process is local.

Signed-off-by: Willy Tarreau <w@1wt.eu>