staging: slhttp: support sending multiple packet responses
We now support sending up to 5 extra data packets in keep-alive mode
and 7 extra packets in close mode. In this situation, the first packet
is the only one which cannot be full, and the other ones are all set
to 1457 bytes exactly (91*16 + 1) so that each packet increases the
state by one.

Eventhough these packets hold a PUSH, it appears that the client does
not ACK them all immediately (40ms delayed ACK). We need to improve
this to try to send the packets 2 at a time.

Signed-off-by: Willy Tarreau <w@1wt.eu>